[any message]

+++ src/main/java/Main.java
@@ -19,7 +19,7 @@ public class Main {
        serverSocket.setReuseAddress(true);
        // Wait for connection from client.
        clientSocket = serverSocket.accept();
-       long message_size = 0, corelation_id = 7;
+       int message_size = 0, corelation_id = 7;
        clientSocket.getOutputStream().write(String.valueOf(message_size).getBytes());
        clientSocket.getOutputStream().write(String.valueOf(corelation_id).getBytes());
        clientSocket.getOutputStream().flush();
